Merge: Attack in totalScene

+++ Assets/Sources/Actions/Fragments/StopPlayerStateFragment.cs
@@ -11,15 +11,15 @@ namespace NotionWorld.Actions
 {
     public sealed class StopAutoAttackFragment : Fragment
     {
-        public int SkillInternal;
+        public float SkillInternal;
 
-        public StopAutoAttackFragment(int skillInternal)
+        public StopAutoAttackFragment(float skillInternal)
         {
             SkillInternal = skillInternal;
         }
         public override void TakeEffect(Entity actor)
         {
-            Delay(SkillInternal * 1000 /5, actor);
+            Delay((int)(SkillInternal * 1000 /5), actor);
         }
         private async void Delay(int ms, Entity actor)
         {
